Updated module augmentation for enter package.
