+++ packages/ckeditor5-enter/src/index.ts
@@ -12,3 +12,5 @@ export { default as ShiftEnter } from './shiftenter';
 export type { ViewDocumentEnterEvent } from './enterobserver';
 export type { default as EnterCommand } from './entercommand';
 export type { default as ShiftEnterCommand } from './shiftentercommand';
+
+import './augmentation';
