+++ packages/ckeditor5-enter/src/shiftentercommand.ts
@@ -156,9 +156,3 @@ function isInsideLimitElement( element: Element, schema: Schema ): boolean {
 
 	return schema.isLimit( element ) || isInsideLimitElement( element.parent as Element, schema );
 }
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface CommandsMap {
-		shiftEnter: ShiftEnterCommand;
-	}
-}
